requirements.txt
gradio
llama-cpp-python
sqlite3
pandas